examples: rename regmap_simp_fun → regmap (AXI4-Lite control pattern)
Pattern-name the directory (the dir names the control pattern; the simple-
function computation keeps its identity in the files/symbols inside —
simp_fun.py, SimpFunComponent, build_simp_fun_dag, and the pysilicon_simp_fun_proj
Vitis project are unchanged).

git mv examples/regmap_simp_fun → examples/regmap and
docs/examples/regmap_simp_fun → docs/examples/regmap; update the build-script and
test imports (examples.regmap_simp_fun.* → examples.regmap.*), the doc-page links
and front-matter (title → "Register Map (simple function)", teaching-order
nav_order, child parent:), the regmap guide link, and the README/comment strings.
