+++ examples/regmap/include/streamutils.cpp
