+++ examples/regmap/include/streamutils_hls.h
