+++ examples/regmap/run.tcl
