+++ examples/regmap/simp_fun.py
